# Replace mocha-phantomjs with mocha-phantomjs-core