This means that PhantomJS isn't installed alongside Shunter, which speeds up npm install times quite a lot. Relates to #49.