@rowanmanning 2015-09-25
I ran some before/after benchmarks on `npm install` for this work. All of these were run with a primed cache and values are in seconds.

With PhantomJS pre-downloaded (so it's just grabbed from a cached zip file):

```
Before:  27.75s
After:   30.07s  (2.32s saving)
```

With PhantomJS download time included:

```
Before:  27.75s
After:   38.96s  (11.21s saving)
```

Timings were calculated by taking the average from five runs of:

```sh
rm -rf node_modules && time npm install
```